Refactor_2024.12.08.12.34_MatrixRowコンポーネントのリファクタリング
